% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linearOP}
\alias{linearOP}
\title{linearOP}
\usage{
linearOP(x, data, penalty, cc = FALSE)
}
\arguments{
\item{x}{a vector (see data)}

\item{data}{a vector defining the data points (x[i], data[i])}

\item{penalty}{the penalty for introducing a new segment}

\item{cc}{a boolean to impose a continuity constraint}
}
\description{
An optimal partitioning algorithm with a linear fit for each segment
}
