% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeOP.R
\name{plot.slopeOP}
\alias{plot.slopeOP}
\title{plot.slopeOP}
\usage{
\method{plot}{slopeOP}(x, ..., data, chpt = NULL, states = NULL)
}
\arguments{
\item{x}{a slopeOP class object}

\item{...}{Other parameters}

\item{data}{the data from which we get res}

\item{chpt}{vector of changepoints of the model}

\item{states}{vector of states of the model}
}
\value{
plot data and the inferred slopeOP result
}
\description{
Plot the result of the slopeOP function with the data
}
