% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_estimation.R
\name{sdHallDiff}
\alias{sdHallDiff}
\title{sdHallDiff}
\usage{
sdHallDiff(data)
}
\arguments{
\item{data}{vector of data to segment: a univariate time series}
}
\value{
an estimation of the sd
}
\description{
Estimation of the standard deviation using the HallDiff estimator
}
\examples{
myData <- slopeData(index = c(1,100,200,300), states = c(0,5,3,6), noise = 1)
sdHallDiff(data = myData)
}
