% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeOP.R
\name{slopeData}
\alias{slopeData}
\title{slopeData}
\usage{
slopeData(index = c(0), states = c(0), noise = 0)
}
\arguments{
\item{index}{a vector of increasing changepoint indexes}

\item{states}{vector of successive states}

\item{noise}{noise level = standard deviation of an additional normal noise}
}
\value{
a vector of simulated data
}
\description{
Generate data with a given continuous piecewise linear model
}
