% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeOP.R
\name{slopeOP}
\alias{slopeOP}
\title{slopeOP}
\usage{
slopeOP(data = c(0), states = c(0), penalty = 0, type = "null")
}
\arguments{
\item{data}{vector of data to segment}

\item{states}{vector of states = accessible ending values}

\item{penalty}{the penalty coefficient. A positive number}

\item{type}{string defining the pruning type to use}
}
\value{
a list of two vectors and a value  = (changepoints, state parameters, global cost)
'changepoints' is the vector of changepoints (we give the last element of each segment (+ 1 the starting index))
'states' is the vector of successive final values of each segment (+ the very first value)
'globalCost' is a number equal to the global cost of the penalized changepoint problem
}
\description{
Optimal partitioning algorithm for change-in-slope problem with a finite number of states (initial and final values of each segment is restricted to a finite set of values).
The algorithm takes into account a continuity constraint between successive segments
}
