% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeOP.R
\name{slopeData}
\alias{slopeData}
\title{slopeData}
\usage{
slopeData(index, states, noise = 0, outlierDensity = 0,
  outlierNoise = 50)
}
\arguments{
\item{index}{a vector of increasing changepoint indices}

\item{states}{vector of successive states}

\item{noise}{noise level = standard deviation of an additional normal noise}

\item{outlierDensity}{probability for a datapoint to be an outlier (has to be close to 0)}

\item{outlierNoise}{noise level for outlier data points}
}
\value{
a vector of simulated data
}
\description{
Generate data with a given continuous piecewise linear model
}
\examples{
myData <- slopeData(index = c(1,100,200,300), states = c(0,5,3,6), noise = 1)
}
